Implement parsing for the records
As agreed in LDM, the specification requires that a data contextual
keyword be present in front of the class and a parameter list to be
present after the identifier, for the type to be a record.

The absence of either of these two elements will be a semantic error,
not a parsing error.